+++ CHANGELOG.md
@@ -11,3 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial repository setup and Go module initialisation.
 - Added Smoke Test for immediate CI/CD pipeline testing
 - Github Actions CI/CD pipeline with `golangci-lint` and strict race detection.
+
+## [0.1.0] - 2026-02-16
+
+### Added
+
+- Installed go-fiber for the backend API
+- Setup a simple backend API
+- Embedded index html file and the static files
+- Added htmx import tag to index.html
+- initalised the frontend serving